JWT 토큰 인가 완료(Board delete만)

+++ src/main/java/site/metacoding/white/service/BoardService.java
@@ -76,12 +76,17 @@ public class BoardService {
     }
 
     @Transactional
-    public void deleteById(Long id) {
+    public void deleteById(Long id, Long userId) {
         Optional<Board> boardOP = boardRepository.findById(id);
+
         if (boardOP.isPresent()) {
+            Board boardPS = boardOP.get();
+            if (boardPS.getUser().getId() != userId) {
+                throw new RuntimeException("해당 글을 삭제할 권한이 없습니다.");
+            }
             boardRepository.deleteById(id);
         } else {
-            throw new RuntimeException("해당" + id + "로 삭제를 할 수 없습니다.");
+            throw new RuntimeException("해당 " + id + "로 삭제를 할 수 없습니다.");
         }
     }
 }
